Move premium status into a single reusable control
