--- FluentFlyoutWPF/Controls/PremiumPurchaseButton.xaml
@@ -1,14 +1,0 @@
-<UserControl x:Class="FluentFlyout.Controls.PremiumPurchaseButton"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
-             x:Name="Root">
-    <ui:Button Click="PurchaseButton_Click"
-               Appearance="Primary"
-               FontSize="12">
-        <StackPanel Orientation="Vertical">
-            <TextBlock Text="{Binding UserSettings.PremiumPurchaseAction, ElementName=Root}" FontWeight="SemiBold" />
-            <TextBlock Text="{Binding UserSettings.PremiumPrice, ElementName=Root}" />
-        </StackPanel>
-    </ui:Button>
-</UserControl>
