--- FluentFlyoutWPF/Controls/PremiumPurchaseButton.xaml.cs
@@ -1,25 +1,0 @@
-// Copyright (c) 2024-2026 The FluentFlyout Authors
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-using FluentFlyout.Classes;
-using FluentFlyout.Classes.Settings;
-using FluentFlyoutWPF.ViewModels;
-using System.Windows;
-using System.Windows.Controls;
-
-namespace FluentFlyout.Controls;
-
-public partial class PremiumPurchaseButton : UserControl
-{
-    public UserSettings UserSettings => SettingsManager.Current;
-
-    public PremiumPurchaseButton()
-    {
-        InitializeComponent();
-    }
-
-    private void PurchaseButton_Click(object sender, RoutedEventArgs e)
-    {
-        LicenseManager.UnlockPremium(sender);
-    }
-}
