+++ FluentFlyoutWPF/Windows/OnboardingWindow.xaml
@@ -162,47 +162,7 @@
                                     <ui:TextBlock FontSize="14" Margin="12,0,12,0" Opacity="0.7" TextWrapping="WrapWithOverflow"
                                         Text="{Binding Source={StaticResource AllFuturePremiumFeatures}, StringFormat='•  {0}'}" />
 
-                                    <!-- Premium Status InfoBar (only shown for Store versions) -->
-                                    <ui:InfoBar Grid.Row="1" Title="{DynamicResource PremiumUnlockedMessage}" Severity="Success" IsClosable="True" Margin="0,8,0,0">
-                                        <ui:InfoBar.Style>
-                                            <Style TargetType="ui:InfoBar" BasedOn="{StaticResource {x:Type ui:InfoBar}}">
-                                                <Setter Property="IsOpen" Value="False" />
-                                                <Style.Triggers>
-                                                    <MultiDataTrigger>
-                                                        <MultiDataTrigger.Conditions>
-                                                            <Condition Binding="{Binding IsStoreVersion, Source={x:Static settings:SettingsManager.Current}}" Value="True"/>
-                                                            <Condition Binding="{Binding IsPremiumUnlocked, Source={x:Static settings:SettingsManager.Current}}" Value="True"/>
-                                                        </MultiDataTrigger.Conditions>
-                                                        <Setter Property="IsOpen" Value="True"/>
-                                                    </MultiDataTrigger>
-                                                </Style.Triggers>
-                                            </Style>
-                                        </ui:InfoBar.Style>
-                                    </ui:InfoBar>
-                            
-                                    <!-- Premium Locked InfoBar (only shown for Store versions without premium) -->
-                                    <ui:InfoBar Grid.Row="1" IsOpen="True" Severity="Warning" IsClosable="False" Margin="0,8,0,0">
-                                        <ui:InfoBar.Style>
-                                            <Style TargetType="ui:InfoBar">
-                                                <Setter Property="Visibility" Value="Collapsed"/>
-                                                <Style.Triggers>
-                                                    <MultiDataTrigger>
-                                                        <MultiDataTrigger.Conditions>
-                                                            <Condition Binding="{Binding IsStoreVersion, Source={x:Static settings:SettingsManager.Current}}" Value="True"/>
-                                                            <Condition Binding="{Binding IsPremiumUnlocked, Source={x:Static settings:SettingsManager.Current}}" Value="False"/>
-                                                        </MultiDataTrigger.Conditions>
-                                                        <Setter Property="Visibility" Value="Visible"/>
-                                                    </MultiDataTrigger>
-                                                </Style.Triggers>
-                                            </Style>
-                                        </ui:InfoBar.Style>
-                                        <StackPanel Orientation="Horizontal">
-                                            <fluentflyoutcontrols:PremiumPurchaseButton Margin="4,8,0,8" />
-                                        </StackPanel>
-                                    </ui:InfoBar>
-
-                                    <!-- GitHub Version InfoBar -->
-                                    <ui:InfoBar Grid.Row="1" Title="{DynamicResource PremiumGitHubMessage}" IsOpen="True" Severity="Informational" IsClosable="False" Visibility="{Binding IsStoreVersion, Source={x:Static settings:SettingsManager.Current}, Converter={StaticResource BoolToCollapsedVisibleConverter}}" Margin="0,8,0,0" />
+                                    <fluentflyoutcontrols:PremiumStatus Margin="0,8,0,0" />
                                 </StackPanel>
                             </StackPanel>
                         </ui:PassiveScrollViewer>
